// src/models/UserCreateModel.ts
export type UserCreateModel = {
    /**
     * The user role if you created user.
     */
    role: string;
    /**
     * The user first name if you created user.
     */
    first_name: string;
    /**
     * The user last name if you created user.
     */
    last_name: string;
    /**
     * The user email if you created user.
     */
    email: string;
    /**
     * The user phone if you created user.
     */
    phone: string;
    /**
     * The user birthdate if you created user.
     */
    birthdate: string;
    /**
     * The user password if you created user.
     */
    password: string;
}